+++ CHANGES.md
@@ -5,6 +5,7 @@ Changes
 - 1.6.2
     - UPGRADE: upgrade to Electron 11.0.1
     - IMPROVEMENT: take screen scale factor (pixels vs. dots) into account for native window resizing
+    - IMPROVEMENT: do not hard-code stream video size to 1920x1080 and instead figure it out for "source size" button
 
 - 1.6.1
     - IMPROVEMENT: raise a fatal error message in the UI in case ffmpeg crashes and a system one is not available